\name{FunctSNP-internal}
\title{Internal FunctSNP objects}
\alias{.availableDBs}
\alias{.getFunctSNPrecords}
\alias{.getFunctSNPinput}
\alias{.install.location}
\alias{.localDBs}
\alias{.updateSupportedConfig}
\alias{.valid.speciesCode}
\description{Internal FunctSNP objects.}
\details{These are not to be called by the user.}
\keyword{internal}


